Docker test script
Runs in privileged container because soon I'll be creating new
namespaces, and I want to start in the initial user namespace. This is
the case even though these namespaces will be created by an unprivileged
user.

The test script assumes that docker is runnable as an unprivileged user.
I use docker machine so this is indeed the case.